requirements.txt
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
groq>=0.4.0
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.24.0
datasets>=2.14.0
requests>=2.31.0
